Started work on the graphical menu
